break out base-model.ts from metamodel.ts
